ai-chatbot/requirements.txt
vanna[chromadb]>=2.0.0
groq>=0.5.0
psycopg2-binary>=2.9.9
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
PyJWT>=2.8.0
python-dotenv>=1.0.1
pandas>=2.2.0
